refactor(ir-core,ir): split IR schema into responsibility modules
