+++ packages/ir-core/src/types/diagnostic.ts
@@ -1,0 +1,11 @@
+export interface DiagnosticIR {
+  level: "error" | "warn" | "info";
+  code: string;
+  message: string;
+  source?: {
+    file: string;
+    line?: number;
+    column?: number;
+    viaSourceMap?: boolean;
+  };
+}
