+++ packages/ir-core/src/types/handler.ts
@@ -1,0 +1,18 @@
+export type HandlerResponseMode =
+  | "return"
+  | "response-object"
+  | "next-callback"
+  | "unknown";
+
+export interface HandlerIR {
+  id: string;
+  params: Array<{
+    name: string;
+    role: "request" | "response" | "next" | "custom";
+  }>;
+  bodyRef?: string;
+  async: boolean;
+  semantics?: {
+    responseMode: HandlerResponseMode;
+  };
+}
