+++ packages/ir-core/src/types/http.ts
@@ -1,0 +1,1 @@
+export type HttpMethod = "GET" | "POST" | "PUT" | "DELETE" | "PATCH";
