+++ packages/ir-core/src/types/index.ts
@@ -1,0 +1,6 @@
+export * from "./http.js";
+export * from "./module.js";
+export * from "./route.js";
+export * from "./handler.js";
+export * from "./diagnostic.js";
+export * from "./program.js";
