+++ packages/ir-core/src/types/module.ts
@@ -1,0 +1,6 @@
+export interface ModuleIR {
+  id: string;
+  sourcePath: string;
+  exports: string[];
+  imports: Array<{ spec: string; kind: "esm" | "cjs"; resolved?: string }>;
+}
