+++ packages/ir-core/src/types/route.ts
@@ -1,0 +1,8 @@
+import type { HttpMethod } from "./http.js";
+
+export interface RouteIR {
+  method: HttpMethod;
+  path: string;
+  handlerRef: string;
+  middlewareRefs?: string[];
+}
