+++ packages/ir/src/index.ts
@@ -1,10 +1,1 @@
-export interface RouteIR {
-  method: "GET" | "POST" | "PUT" | "DELETE";
-  path: string;
-  handler: string;
-}
-
-export interface ProgramIR {
-  routes: RouteIR[];
-  warnings: string[];
-}
+export * from "./types/index.js";
