+++ packages/ir/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./route.js";
+export * from "./program.js";
