+++ packages/ir/src/types/program.ts
@@ -1,0 +1,6 @@
+import type { RouteIR } from "./route.js";
+
+export interface ProgramIR {
+  routes: RouteIR[];
+  warnings: string[];
+}
