+++ packages/ir/src/types/route.ts
@@ -1,0 +1,5 @@
+export interface RouteIR {
+  method: "GET" | "POST" | "PUT" | "DELETE";
+  path: string;
+  handler: string;
+}
